new command

+++ commands/insert.js
@@ -5,20 +5,26 @@ module.exports = {
 	description: 'set gifs!',
 	execute: async (msg, args) =>{
 		var functions = require('../functions/functions.js');
-		if(args.lenght)
+		var commands = require('./command.js');
 		var col = args[0];
+		command = await commands.commands(msg,col);
 		var url = args[1];
-		if(msg.member.id==DEV){
+		if(command.description){
+			if(msg.member.id==DEV){
 			if (!args || args == "") {
 				msg.channel.send("I need an url");
 			}	
 			else{
 				await functions.insert(msg,url,col);
 			}
+			}
+			else{
+				msg.channel.send("You dont have permission");
+			}
+			msg.delete();
 		}
 		else{
-			msg.channel.send("You dont have permission");
+			msg.channel.send(`Command **`+col+`** doesn't exist`);
 		}
-		msg.delete();
 	},
 };
